drivechain code (still needs values modified after testing)
